Get mean comparisons from check_model.fit_model_anova object

mean_comparisons performs mean comparisons from object coming from check_model.fit_model_anova

# S3 method for check_model_anova
mean_comparisons(x, alpha = 0.05,
  p.adj = "none", ...)

Arguments

x
outputs from check_model.fit_model_anova

alpha
level of type one error. 0.05 (5%) by default

p.adj
For all except type = 2. NULL for no adjustement of the type one error. p.adj can "holm", "hochberg", "bonferroni", "BH", "BY" or "fdr" p-adj = "none" is t-student. See p.adjust() for more details.

...
further arguments passed to or from other methods

Value

A list of four elements :

info : a list with variable

Details

S3 method. See in the book for more details here

See also