Plot hedonic object from format_data_PPBstats.data_organo_hedonic

plot.data_hedonic gets ggplot to describe the data

# S3 method for data_organo_hedonic
plot(x, plot_type = "boxplot",
  x_axis = NULL, in_col = NULL, vec_variables = NULL,
  nb_parameters_per_plot_x_axis = 5, nb_parameters_per_plot_in_col = 5,
  labels_on = NULL, labels_size = 4, pie_size = 0.2, zoom = 6, ...)

Arguments

x
The data frame coming from format_data_PPBstats.data_organo_hedonic

plot_type
the type of plot you wish. It can be :

"pam" for presence abscence matrix that represent the combinaison of germplasm x location

"histogramm"

"barplot", where sd error are displayed

"boxplot"

"interaction"

"biplot"

"radar"

"raster"

"map"

x_axis
factor displayed on the x.axis of a plot. "date_julian" can be choosen: it will display julian day for a given variable automatically calculated from format_data_PPBstats(). This is possible only for plot_type = "histogramm", "barplot", "boxplot" and "interaction".

in_col
factor displayed in color of a plot

vec_variables
vector of variables to display

nb_parameters_per_plot_x_axis
the number of parameters per plot on x_axis arguments

nb_parameters_per_plot_in_col
the number of parameters per plot for in_col arguments

labels_on
factor to display for plot_type = "biplot"

labels_size
size of the label for plot_type = "biplot" and "radar"

pie_size
when plot_type = "map" and vec_variables is not NULL, size of the pie

zoom
zoom of the map, see ?get_map for more details

...
further arguments passed to or from other methods

Value

For plot_type "histogramm", "barplot", "boxplot" or "interaction", the function returns a list with ggplot objects for each variable of vec_variables.

For plot_type "biplot", the function returns a list with ggplot objects for each pairs of variables of vec_variables.

For plot_type "radar" and "raster, the function returns a list with ggplot objects with all variables of vec_variables.

For plot_type "map", it returns a map with location if vec_variables = NULL and labels_on = "location". If vec_variables is not NULL, it displays pie on map.

Each list is divided in several lists according to values of nb_parameters_per_plot_x_axis and nb_parameters_per_plot_in_col except for plot_type = "map".

Details

See the book for more details : https://priviere.github.io/PPBstats_book/hedonic.html#describe-the-data-2

See also