Get ggplot to visualize output from mean_comparisons.check_model_bh_intra_location

plot.mean_comparisons_model_bh_intra_location returns ggplot to visualize outputs from mean_comparisons.check_model_bh_intra_location

# S3 method for mean_comparisons_model_bh_intra_location
plot(x,
  plot_type = c("interaction", "barplot", "score"),
  nb_parameters_per_plot = 8, ...)

Arguments

x
Output from mean_comparisons.check_model_bh_intra_location

plot_type
"interaction", "barplot" or "score"

nb_parameters_per_plot
number of parameter per plot to display

...
further arguments passed to or from other methods

Value

A list with ggplot object depending on plot_type. For each plot_type, it is a list of three elements being lists with as many elements as environment. For each element of the list, there are as many graph as needed with nb_parameters_per_plot parameters per graph.

barplot :

data_mean_comparisons : only environments where all MCMC converge are represented. Letters are displayed on each bar. Parameters that do not share the same letters are different regarding type I error (alpha) and alpha correction. The error I (alpha) and the alpha correction are displayed in the title. alpha = Imp means that no differences were possible to find.

data_env_with_no_controls : only environments where there were no controls are represented.

data_env_whose_param_did_not_converge : only environments where MCMC did not converge are represented.

interaction :

data_mean_comparisons : only environments where all MCMC converge are represented. The error I (alpha) and the alpha correction are displayed in the x.axis under the form "alpha | alpha correction". alpha = Imp means that no differences were possible to find.

data_env_with_no_controls : only environments where there were no controls are represented.

data_env_whose_param_did_not_converge : only environments where MCMC did not converge are represented.

score : The score is set according to which group the entry was allocated. An high score means that the entry was in a group with an high mean. A low score means that the entry was in a group with an low mean. In the legend, the score goes from 1 (first group) to the number of groups of significativity. The error I (alpha) and the alpha correction are displayed in the x.axis under the form "alpha | alpha correction". alpha = Imp means that no differences were possible to find.

Details

S3 method. See example in the book: https://priviere.github.io/PPBstats_book/family-2.html#model-1

See also